refactor: split profile cli modules
